Automatically clone raft when the raft pinned tag changes (#2087)
Now that RAFT has official conda packages, it'll be much easier to work locally on cugraph features that rely on the RAFT nightly. 

Working on features that span raft and cugraph should be fairly straightforward to do locally as well, because CPM provides the `CPM_raft_SOURCE` variable that can passed into the `cmake` call to tell the build to use a local working copy of RAFT.

The raft conda packages are now ready to add to the CI conda dependencies and once added, the updating the raft pinned tag in `get_raft.cmake` will no longer work. The option added in this PR will force clone a local version of the pinned tag (both locally in CI) and set the `CPM_raft_SOURCE` to that path.

Authors:
  - Corey J. Nolet (https://github.com/cjnolet)

Approvers:
  - AJ Schmidt (https://github.com/ajschmidt8)
  - Chuck Hastings (https://github.com/ChuckHastings)
  - Rick Ratzel (https://github.com/rlratzel)

URL: https://github.com/rapidsai/cugraph/pull/2087