Format Python code with black (#2778)
closes https://github.com/rapidsai/cugraph/issues/2527

I also added `.pre-commit-config.yaml` to make this easier to use and match versions (for example, I kept using a newer flake8 locally).  This uses the versions of black and flake8 that are in the `rapids` environment in the docker container.  There's a lot more we can do with pre-commit, but I didn't want to be overzealous in this PR.

_After_ this PR is merged, we should add a `.git-blame-ignore-revs` file so that I won't get blamed for everything :) .  See:
- https://black.readthedocs.io/en/stable/guides/introducing_black_to_your_project.html#avoiding-ruining-git-blame
- https://docs.github.com/en/repositories/working-with-files/using-files/viewing-a-file#ignore-commits-in-the-blame-view

Once this is merged, existing PRs should probably run `black` (version 22.3.0) on their branch before merging 22.12 branch into their branch.

**This changes the maximum line length to 88, which is the default for black**

Authors:
  - Erik Welch (https://github.com/eriknw)

Approvers:
  - Rick Ratzel (https://github.com/rlratzel)
  - AJ Schmidt (https://github.com/ajschmidt8)

URL: https://github.com/rapidsai/cugraph/pull/2778