Use new `rapids-dask-dependency` metapackage for managing `dask` versions (#3991)
Currently dask versions are pinned as part of every release cycle and then unpinned for the next development cycle across all of RAPIDS. This introduces a great deal of churn. To centralize the dependency, we have created a metapackage to manage the required dask version and this PR introduces that metapackage as a dependency of cugraph.

xref: https://github.com/rapidsai/cudf/pull/14364

Authors:
   - Chuck Hastings (https://github.com/ChuckHastings)
   - Rick Ratzel (https://github.com/rlratzel)
   - Vyas Ramasubramani (https://github.com/vyasr)
   - GALI PREM SAGAR (https://github.com/galipremsagar)
   - Naim (naim@uib.no)

Approvers:
   - Jake Awe (https://github.com/AyodeAwe)