Faster `shortest_path`
For larger graphs, nearly all the time is spent creating dicts of lists of paths.
This PR delays creating the lists of paths by duck-typing the return object
and returning a Mapping that computes paths as needed. There is virtually no
performance penalty for this, and a huge performance gain by delaying computation.