Includes all deferred conversion costs in benchmarks (#34)
This refactors the helper that returns the callable to benchmark to fully expand Mapping return types. This is needed in case those return types lazily convert values from device to host, so the complete cost can be included in the measured/benchmarked run.  This is similar to exhausting returned generators that deferred computations in NetworkX, and doing this keeps us consistent with benchmarking the full compute and/or conversion costs for both libraries.

I changed the option name from `exhaust_returned_iterator` to `force_unlazy_eval` since this is no longer being used only for iterators, but I'm taking suggestions for a better name.

Authors:
  - Rick Ratzel (https://github.com/rlratzel)
  - Erik Welch (https://github.com/eriknw)

Approvers:
  - Erik Welch (https://github.com/eriknw)
  - Ralph Liu (https://github.com/nv-rliu)

URL: https://github.com/rapidsai/nx-cugraph/pull/34