Add `lowest_common_ancestor` algorithm (#35)
This is a medium-priority algorithm that I considered pairing with @nv-rliu on instead of https://github.com/rapidsai/nx-cugraph/pull/32, so I thought I would knock it out while it was still fresh in mind. It was pretty quick.

To reviewers: this is low priority and may slip to the next release. Please do more important things first. It is reasonable to request benchmarking, which I hope to add when I get a chance.

I created more tests in networkx to help with this PR: https://github.com/networkx/networkx/pull/7726

I think the implementation is pretty clean and straightforward, so hopefully others think so too (although it probably requires understanding _what_ this algorithm does).

Authors:
  - Erik Welch (https://github.com/eriknw)

Approvers:
  - Ralph Liu (https://github.com/nv-rliu)
  - Rick Ratzel (https://github.com/rlratzel)

URL: https://github.com/rapidsai/nx-cugraph/pull/35