Fix edgecase when converting to networkx w/o fallback (#118)
This fixes a (probably rare) case that I encountered when calling a function with `nxcg.Graph` and `backend="networkx"` when `nx.config.fallback_to_nx is False` and the function calls another networkx function.

The `nxcg.Graph` is networkx-compatible, but it still has `__networkx_backend__ == "cugraph"`, so there are times when some behaviors may be different (but technically correct) b/c `__networkx_backend__` isn't `"networkx"`. So, we convert fully to networkx graph. An alternative would be to create a proxy graph that shares the data structures of the `nxcg.Graph` but has `G.__networkx_backend__ == "networkx"`, but the current approach is simplest and adequately covers what should be a very rare occurance.

Regression test added.

Authors:
  - Erik Welch (https://github.com/eriknw)

Approvers:
  - Rick Ratzel (https://github.com/rlratzel)

URL: https://github.com/rapidsai/nx-cugraph/pull/118